requirements.txt
addfips==0.2.2
aiohttp==3.6.2
appdirs==1.4.4
appnope==0.1.0
area==1.1.1
arrow==1.0.3
async-timeout==3.0.1
asyncio==3.4.3
autosig==0.10.0
backcall==0.1.0
beautifulsoup4==4.8.2
binaryornot==0.4.4
bleach==3.1.0
cachetools==4.1.1
census==0.8.17
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
cookiecutter==1.7.2
csv-diff==0.6
cycler==0.10.0
decompose-mrobison==0.0.2
decorator==4.4.0
defusedxml==0.6.0
dictdiffer==0.8.1
distlib==0.3.1
dnspython==1.16.0
docopt==0.6.2
docxtpl==0.10.0
Elixir==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
falcon==2.0.0
ffmpeg==1.4
filelock==3.0.12
Fiona==1.8.14
flake8==3.8.1
Flask==1.1.1
flatten-json==0.1.13
geojson==2.5.0
geopandas==0.8.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
haikunator==2.1.0
halo==0.0.30
html5lib==1.1
httplib2==0.18.1
hug==2.6.1
idna==2.10
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jellyfish==0.6.1
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.2.0
log-symbols==0.0.14
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.20.0
more-itertools==8.4.0
multidict==4.7.6
munch==2.5.0
nameparser==1.0.6
names==0.3.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
num2words==0.5.10
numpy==1.19.0
oauthlib==3.1.0
openpyxl==3.0.4
packaging==20.4
pandas==1.2.4
pandocfilters==1.4.2
parso==0.5.1
pbr==5.4.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
poyo==0.5.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.4.2
pygsheets==2.0.5
pymongo==3.10.1
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.15.4
pytesseract==0.3.7
pytest==5.4.3
python-box==5.3.0
python-dateutil==2.8.0
python-docx==0.8.10
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.5.4
remote-pdb==2.1.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0
spinners==0.0.24
SQLAlchemy==1.3.13
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
text2num==2.2.1
tika==1.24
toml==0.10.0
toolz==0.10.0
tornado==6.0.3
tox==3.15.0
traitlets==4.3.2
typing==3.7.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
us==2.0.2
virtualenv==20.0.35
wcwidth==0.2.5
webencodings==0.5.1
Webflowpy==0.0.4
Werkzeug==0.15.6
widgetsnbextension==3.5.1
words2num==0.2.3
xlrd==1.2.0
xlwt==0.7.2
yarl==1.4.2
